Moved convertRatingToStars function to /lib
