+++ client/src/assets/ratings-and-reviews.sass
@@ -35,6 +35,8 @@
           #averageRatingStars
             display: flex
             flex-direction: row
+            margin-top: 10px
+            margin-left: 9px
 
             .rating
               display: none
