+++ client/src/components/RatingsAndReviews/RatingBreakdown.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import StarRating from './StarRating.jsx';
-import { getAverageRating, calculateTotalReviews, calculateRatingsPercentage, getRecommendPercentage } from '../../lib/ratingsAndReviewsHelpers.js';
+import { getAverageRating, calculateTotalReviews, calculateRatingsPercentage, getRecommendPercentage, convertRatingToStars } from '../../lib/ratingsAndReviewsHelpers.js';
 import { getReviewsMeta } from '../../lib/requestHelpers.js';
 
 
 
-const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeStars, removeAllFilters, convertRatingToStars }) => {
+const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeStars, removeAllFilters }) => {
 
   const [filterActive, setFilterActive] = useState(false);
 
@@ -19,7 +19,9 @@ const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeSta
     <div id="breakdown">
       <div id="averageRatingHeading">
         <div id="averageRatingNumber">{averageRating}</div>
-        {convertRatingToStars(averageRating)}
+        <div id="averageRatingStars">
+          {convertRatingToStars(averageRating)}
+        </div>
       </div>
       <div onClick={() => {
         sortReviews(5);
